// generated/models/ApiCompetitionDraftInput.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiCompetitionRulesInput } from '../models/ApiCompetitionRulesInput';
import { ApiCreateNewWaveParticipationConfig } from '../models/ApiCreateNewWaveParticipationConfig';
import { ApiCreateNewWaveVotingConfig } from '../models/ApiCreateNewWaveVotingConfig';
import { ApiCreateWaveMetadataRequest } from '../models/ApiCreateWaveMetadataRequest';
import { ApiCreateWaveOutcome } from '../models/ApiCreateWaveOutcome';
import { HttpFile } from '../http/http';

export class ApiCompetitionDraftInput {
    'title': string;
    'description': string | null;
    'participation': ApiCreateNewWaveParticipationConfig;
    'voting': ApiCreateNewWaveVotingConfig;
    'rules': ApiCompetitionRulesInput;
    'outcomes': Array<ApiCreateWaveOutcome>;
    'presentation': Array<ApiCreateWaveMetadataRequest>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "participation",
            "baseName": "participation",
            "type": "ApiCreateNewWaveParticipationConfig",
            "format": ""
        },
        {
            "name": "voting",
            "baseName": "voting",
            "type": "ApiCreateNewWaveVotingConfig",
            "format": ""
        },
        {
            "name": "rules",
            "baseName": "rules",
            "type": "ApiCompetitionRulesInput",
            "format": ""
        },
        {
            "name": "outcomes",
            "baseName": "outcomes",
            "type": "Array<ApiCreateWaveOutcome>",
            "format": ""
        },
        {
            "name": "presentation",
            "baseName": "presentation",
            "type": "Array<ApiCreateWaveMetadataRequest>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCompetitionDraftInput.attributeTypeMap;
    }

    public constructor() {
    }
}